Use bindings.yaml file in tests.

+++ test/headers/cpp/default_values.hpp
@@ -1,5 +1,3 @@
-#include <cstddef>
-
 namespace cv
 {
   class Range
@@ -23,7 +21,7 @@ namespace cv
     Mat(const Mat& m, const Range& rowRange, const Range& colRange = Range::all());
     // Constructor with anonymous enum default value
     // Should generate: cv::Mat::AUTO_STEP, not cv::Mat::(unnamed enum at ...)::AUTO_STEP
-    Mat(int rows, int cols, void* data, size_t step = AUTO_STEP);
+    Mat(int rows, int cols, void* data, int step = AUTO_STEP);
   };
 
   // Test default values in class templates with nested types
